requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
albumentations>=1.3.0
pillow>=10.0.0

# SAM (Segment Anything)
segment-anything>=1.0

# LangChain & Agent
langchain>=0.1.0
langgraph>=0.0.20

# Data & Config
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0

# Performance monitoring
psutil>=5.9.0

# TensorRT (optional, for deployment)
# tensorrt>=8.6.0
# onnx>=1.14.0
# onnxruntime-gpu>=1.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
